RNN
Language Translation
In this project, you’re going to take a peek into the realm of neural network machine
translation. You’ll be training a sequence to sequence model on a dataset of English and French
sentences that can translate new sentences from English to French.

Build the Neural Network
You'll build the components necessary to build a Sequence-to-Sequence model by
implementing the following functions below:
● model_inputs
● process_decoding_input
● encoding_layer
● decoding_layer_train
● decoding_layer_infer
● decoding_layer
● seq2seq_model

Implement the model_inputs() function to create TF Placeholders for the Neural Network. It should
create the following placeholders:
● Input text placeholder named "input" using the TF Placeholder name parameter
with rank 2.
● Targets placeholder with rank 2.
● Learning rate placeholder with rank 0.
● Keep probability placeholder named "keep_prob" using the TF Placeholder name
parameter with rank 0.
Return the placeholders in the following the tuple (Input, Targets, Learning Rate, Keep Probability)